+++ Runtime/EasterAd/EasterAdInteraction.cs
@@ -1,3 +1,4 @@
+#nullable enable annotations
 using System;
 using UnityEngine;